[executor] trivial: rename
