+++ execution/executor/src/components/mod.rs
@@ -4,6 +4,6 @@
 #![forbid(unsafe_code)]
 
 pub mod apply_chunk_output;
+pub mod block_tree;
 pub mod chunk_commit_queue;
 pub mod chunk_output;
-pub mod speculation_cache;
